refactor(bayn): modularize health
